+++ gitlabform/constants.py
@@ -1,0 +1,4 @@
+# f.e. bad syntax in the config file ~= "it's your fault" 😅
+EXIT_INVALID_INPUT = 1
+# f.e. when requests to GitLab fail ~= "it's not your fault" 😎
+EXIT_PROCESSING_ERROR = 2
