+++ gitlabform/processors/util/branch_protector.py
@@ -1,7 +1,7 @@
 from logging import debug
 from cli_ui import warning, fatal
 
-from gitlabform import EXIT_PROCESSING_ERROR, EXIT_INVALID_INPUT
+from gitlabform.constants import EXIT_PROCESSING_ERROR, EXIT_INVALID_INPUT
 from gitlabform.gitlab import GitLab
 from gitlabform.gitlab.core import NotFoundException
