+++ tests/unit/test_transform.py
@@ -7,7 +7,7 @@ from deepdiff import DeepDiff
 from gitlabform import EXIT_INVALID_INPUT
 from gitlabform.configuration import Configuration
 from gitlabform.gitlab import GitLab
-from gitlabform.transform import (
+from gitlabform.configuration.transform import (
     AccessLevelsTransformer,
     UserTransformer,
     ImplicitNameTransformer,
